Prepare the bastion host to deploy the cluter from with ansible.
An ansible playbook and its requirements have been added to prepare the bastion host for the OCP cluster deployment.

Added a variable to terraform to define the public DNS domain within which the private domain for the cluster will be created.  And the

Added a variable to terraform to define the VPC network segment

The external public DNS zone is mantained to resolve the name of the bastion host

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>