Use a random suffix for the ssh key name and remove the input var ssh-keyname.  Default route to nat gateways use the correct nat_gateway_id
A random string of 5 characters is appended to the fix name of the ssh keyname to to make sure that the name created by terraform does not already exist

nat_gateway_id is used instead of gateway_id, that worked by caused the route configuration to be reapplied everytime terraform apply was run.

Add the variable terraform_created to signal whether terraform was used to create the infrastructure or not.  Ansible uses this boolean with when blocks to run or not, terraform output and later load the terraform vars.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>