Use RHEL 8.5 as the base OS for the bastion host
The pool to subscribe the bastion to is defined in a variable.
The rpm packages repositories to add to the bastion are defined in a variable, if this variable is empty no attemp is made to add any repos.
Default DNS hosted zone has been changed, in most cases this needs to be defined via a variable when running terraform.
AMIS  updated to use RHEL 8.5.

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>