The RHEL AMI for the bastion host is obtained based on a filtered query.
The oc client and Opeshift installer are downloaded automatically in the bastion host, no need for the user to do it himself.

The RHEL bastion host does not need to be registered with Red Hat.

Terraform variables are loaded into the ansible playbook

The user need not update the ansible inventory file manually anymore, it is automatically updated for him inside the localhost play

ssh key for the install config file is obtained automatically based on terraform variable ssh_keyfile, the user has no longer need to provide this file for the ansible playbook

Default value for the compute and master instance type in AWS changed to m5.xlarge, since m4 family is not available in all regions

Signed-off-by: Jose Ignacio Jerez Rodriguez <jjerezro@redhat.com>